--- a/pybamm/solvers/c_solvers/__init__.py
+++ b/pybamm/solvers/c_solvers/__init__.py
@@ -1,0 +1,10 @@
+import lazy_loader
+
+
+__getattr__, __dir__, __all__ = lazy_loader.attach(
+    __name__,
+    submodules={},
+    submod_attrs={},
+)
+
+__all__ = []