@agriyakhetarpal 2024-01-20
I might have solved it: for modules that further have sub-modules (`pybamm.models.*`), we might need to modify every `__init__.py` file, not just the base one. I tried [this command](https://github.com/Erotemic/mkinit#command-line-usage) in the root directory which set everything up for me automatically:

```bash
mkinit --lazy_loader --inplace --recursive pybamm  # also accepts --noall so as to not add __all__ in the files
```

and I presumably removed the import caches (all `__pycache__` folders) recursively by doing

```bash
rm -rf $(find . -type d -name '__pycache__')
```

But as I understand, this still does not remove all caches – so `%timeit importlib.import_module("pybamm")` in a interactive REPL might be returning an incorrect calculation and could be unreliable. It does _feel_ faster, though – but it could be a case of placebo.

<details><summary>Outputs from importing PyBaMM </summary>

Without `lazy_loader`

```
%timeit importlib.import_module("pybamm")

The slowest run took 13.77 times longer than the fastest. This could mean that an intermediate result is being cached.
2.33 µs ± 3.05 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

With `lazy_loader`

```
%timeit importlib.import_module("pybamm")

The slowest run took 11.64 times longer than the fastest. This could mean that an intermediate result is being cached.
2.29 µs ± 2.84 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

</details>

This isn't helpful TBH, so I started deleting the entries from the `sys.modules` dictionary:

```python
import sys
keys_to_delete = [key for key in sys.modules if 'pybamm' in key]
for key in keys_to_delete:
    del sys.modules[key]
```

in addition to removing the `__pycache__` directories, but I haven't found a substantial increase yet. Maybe this can help act as a precursor for your experiments as we go further. A reliable method, though a bit tedious, would be to re-clone the repository, re-install it from source, apply the lazy_loader changes, and then benchmark the `import pybamm` statement – maybe try that and see what you get?

P.S. You can consider putting out a testimonial (https://github.com/scientific-python/lazy_loader/issues/50) after we manage to do this